Fix GhostPDF integration with GS - Use of /Page#
We added /Page# to the dictionary returned from pdfgetpage in order to
get pdf2dsc.ps to work (it relies upon that key being present) and
potentially other PostScript programs as well. The key was added to
the dictionary in pdfdopages.

However, after adding it, we then proceeded to assume it would always be
present, even if we called pdfpage and friends directly, rather than
using pdfdopages.

Fix that assumption here by adding the key to the dictionary in
pdfgetpage instead of pdfdopages. There's no way (using the old
PostScript code) to use any of the other functions without using
pdfgetpage first.