GhostPDF - cater for padding in ObjStm
Bug #705259 "Bug with new PDF-Interpreter"

The PDF file here has an ObjStm where the /First value (offset in bytes
from the start of the uncompressed stream to the first object) is not
the same as the nd of the initial table of objects and offsets. There is
padding after the table before the start of the first object.

Obviously this simply makes the stream (slightly) larger and slower to
decompress.

We weren't handling that, we assumed that the fist object lay
immediately after the table. This commit closes the stream after we
have processed the table, because we don't track how many bytes we
consumed reading the table (no way currently to do so) and so cannot
know how many bytes we still need to read.

We then reopen the stream (and any required filters) read and discard
as many bytes as are needed and then read the relevant object.

This is slightly slower (obviously), if necessary we could change the
loop which reads the table to manually read the values instead of using
the functions we already have to do so, which would allow us to track
how many bytes we have read and avoid the open/close. If this ever
looks like its a performance bottleneck we can do that.