GhostPDF - fix memory alignment issues with path code
The deferred path construction stored the segment type as a char
in the same block of memory as the double values holding the points.
This doesn't work on certain processors (ARM is an example, there are others)
which require certain data types to be aligned on boundaries.

To avoid the overhead of up to 7 bytes per segment by aligning the
double pointer I've moved to 2 arrays; one to hold the segment
types and one to hold the co-ordinates. This allows us to keep the
memory overhead zero at the cost of some code complexity.