FirstPage/LastPage device - fix order of processing
Bug #705219 "Using ghostscript with the new PDF engine always only outputs 1 page when using -sPageList"

Not sure why this doesn't cause a problem with regular Ghostscript. The
problem is caused by the order of processing page device parameters in
the device.

We started by reading the 'DisablePageHandler' and potentially disabling
the device (which we do for PDF input, as the interpreter handles this).
But we carried on and processed the remaining parameters, some of which
would then reset DisablePageHandler, effectively re-enabling the device.

This commit simply prevents us processing any of the relevant parameters
if the page handler device is disabled.

This isn't something which gets well tested, so there may be additional
bugs lurking.