txtwrite - improve handling of direct CID-Keyed fonts
Bug #705288 "crash with c stack overflow with sDevice=txtwrite"

The txtwrite device does not currently support the direct use of
CID-Keyed fonts, CIDFonts are only supported as descendants of a Type 0
font.

This caused a couple of problems; firstly the text_resync() method
ended up being recursive and secondly we tried to treat a CID-Keyed
font as if it were a type 0 font, with predictably fatal consequences.

This commit does not add support for CID-Keyed fonts, but it prevents
the infinite recursion. Use of a CID-Keyed font now produces a warning
and exits gracefully with an error.