GS + GhostPDF - only set PageSpotColors for spot-capable devices
Sending PageSpotColors to the txtwrite device was causing it to throw
an error on setpagdevice as it did not process the key. While that
shouldn't really cause a problem (and doesn't with the following commit)
we don't really want to set the key for devices which aren't capable of
producing spots anyway.

This only affects the Ghostscript implementation of GhostPDF.