txtwrite - better error handling and improved file handling
Bug #705289 "Txtwrite outputs only last page of multi-page pdf"

The new PDF interpreter was sending an unrecognised page key to the
device (PageSpotColors) which was causing an error from
gx_default_put_params(). Because of some trickery regarding media size
changes this was leading to us marking the device as closed, and so
each page re-opened it, causing us to open a new file, and therefore
overwriting the existing one.

In addition, the device didn't check to see if a new output filename
was the same as the existing one, but simply closed the existing file
and opened it anew, again overwriting the existing file. I'm not clear
on why this doesn't happen with the old interpreter but it's fairly
clearly a bug.

This commit resolves both of these problems.