txtwrite - update to handle PDF type 3 fonts from new interpreter
Bug #705292 "text output is empty"

The device does in fact state that there was an error processing the
document, so empty output is not a huge surprise.

The new PDF interpreter handles type 3 fonts differently to PostScript
type 3 fonts (because it isn't written in PostScript), and we hadn't
updated the txtwrite devcie to deal with that (pdfwrite was updated).

Although we can handle the type 3 font without exiting the device now,
we can still only deal with a single type 3 character at a time, so we
need to behave similarly to the pdfwrite code. I've updated the
processing loop to do that, and the output from txtwrite is now identical
to that of the old interpreter.

At the same time, resolved a compiler warning from a prior commit.

Finally; add copyright boilerplate to the doc_common.c and .h files
it was apparently not included there previously.....