GS + GhostPDF - close the PDF file when complete
The old code relied upon runpdf closing the file supplied to the /run
(redefined PS operator) function instead of closing the file itself.

For compatibility this now does the same with the new PDF interpreter
code. Note that unless PDFSTOPONERROR is true we will execute the PDF
file in a stopped context, so as to allow us to close the file even in
the event of an error.