GS + GhostPDF - make the circular reference error a standard error
Previously exiting with the error gs_error_circular_reference would
cause the PostScript interpreter to throw a non-standard error and
dump extra information to the back channel. We don't really want that
to happen, we just want to treat it as an error.

This commit moves the error definition into the 'normal' range, rejigs
the PDF interpreter error names table to reflect that, and updates the
PostScript error names to include the circular reference error so that
it is treated like other errors.