GS + GhostPDF - protect against circular references in the Info dict
OSS-fuzz #47491

When retrieving the PDF (and page) information we turn the supplied PDF
dictionary into a PostScript dictionary (and all its contents too).

This isn't a problem for the page dictionary as the interpreter builds
that for us from other information, and it won't contain any circular
references.

For the PDF Info dict, however, we just use the dictionary as stored by
the PDF interpreter, and that won't be guaranteed to be free from
potential circular references so we need to protect against it in the
PostScript interface code. It also means that should we use this code
for other purposes in future it's already protected against this kind of
error.