GhostPDF - improve circular reference detection in /Pages tree
OSS-fuzz #47542.pdf

We didn't have any specific protection against loops in the Pages tree
but simply relied on the dereferencing code to deal with it. This turns
out to be insufficient because we replace indirect references in the
nodes as we go.

So here we add the object numbers of directly defined objects when we
encounter them, mimicking the procedure done when we dereference an
object. This does also mean we have to mark and clear the stored object
numbers back at every level of the page tree.