OSS-fuzz #47522 - typo checking stack in pdfi_repair_file
When checking to see that we have an 'x y obj' we need to check that
both x and y are integers. Due to a typo we were checking y twice.
This could let a non-integer object slip through the checks.

Previously this would have caused us to store an incorrect number but
since the change to storing certain objects as low value integers,
instead of pointers to structures, if the object was a boolean then
we would end up trying to read from low memory locations causing a
seg fault.