oss-fuzz 47536: Handle allocation failure for clip rect list
In the event that we encounter an allocation failure for the rectangle list
in a complex clip path, we'd previously leave the point set to NULL, which
worked well for Postscript where processing would stop on such an error.

With PDF, we have to try to carry on, meaning we'd later try to dereference
that NULL pointer.

So, in that event, put the old rectangle list back (and return the error code).

This reflects what happens for regular paths.