GhostPDF - don't throw errors on unhandled data types converting to PS
When retrieving the Info dictionary for PostScript we turn the content
of the dictionary into PostScript objects. However there are some
objects which we don't handle, specifically streams, and previously we
threw an error if presented with one.

These shouldn't really be present in the Info dictionary, there's
nothing useful you can do with them there, but ignoring the error makes
a couple of odd files work without problems. If we find something we
can't handle we know just make a null object for it, so at least the
user can tell there was a key/value pair in the dictionary and can read
teh key.