GhostXPS - check a colour has enough data before use.
Bug #705333 "Uninitialized read in xps_parse_color"

The code was assuming that if there was too little data for a 4
component colour then it must be a three component colour, but didn't
check to see there was enough to satisfy that.

Check the data, and if there is insufficient, then ignore the colour
(warn in a debug build). We can't return an error as this is a void
function.