GS + GhostPDF - Honour PDFSTOPONERROR
Previously we ran the PDF interpreter from the PostScript interface
more or less totally in a stopped context, which meant that it never
exited with an error status, even when the file had an error.

We now use PDFSTOPONERROR to determine whether the PDF interpreter
should run in a stopped context or should actually return a PostScript
error, which makes the interpreter exit with a non-zero status.

This causes two files, output from pdfwrite, to throw errors where they
previously did not when tested on the cluster. This is actually correct
behaviour when compared with the action of the old PDF interpreter.