GhostPDF - add more warnings for invalid PDF files
Starting with:

Bug #704948 "Warn about xref stream pointing to xref table"

I've chosen for now at least to make this an error, but continue and
read the xref table anyway unless PDFSTOPONERROR is set. In any event
flag the error.

Also check 'stream' keywords to see that they are terminated by a
linefeed as per the specification. We've seen files where a carriage
return on it's own is used, and if we ignore the problem then the
file runs to completion so this is just a warning currently.

And finally; if we encounter standard xref table entries which are not
exactly 20 bytes long, flag a warning. Again we've seen plenty of
examples where the entry is terminated with a linefeed instead of a
carriage return and linefeed or padding space and linefeed, and if we
ignore the error then the file runs properly. So again this is a warning
not an error.