GhostPDF - implement named Destinations
Bug #705218 "table of contents/bookmarks wrong after merging PDF files"

There are actually two problems here, this commit deals with the first
problem.

Destinations for Link annotations and Outlines need not be an array,
they can be a name or string and refer to a named destination. We need
to look up the name in the Dests name tree, which is stored in the
Names dictionary in the Catalog (Root) dictionary.

This deals with the fact that file1.pdf from this bug report has an
Outline with a Named Destination. We still need to deal with offsetting
the page numbers of the Destinations by the number of pages already
processed in previous files.