Bug 705321: More refinement of symbolic TT font with encoding
The spec states that a font marked as symbolic shouldn't have an encoding,
and, for TTF, should include a 3,0 cmap table in the font. The spec does not
cover what to do when those conditions aren't met.

Latest guess: if a symbolic font doesn't have a 3,0 (symbolic) cmap, and *does*
have a 3,1 (Unicode) cmap, then remove the symbolic flag, and use
WinAnsiEncoding. If the font doesn't have a 3,1 cmap, it *should* have a 1,0
(Macintosh) cmap, so use MacRomanEncoding.

Finally, non-symbolic fonts do not draw the TTF notdef, whilest symbolic fonts
do draw it. Include a (slightly horrid) way to cope with that.