GhostPDF - Handle Outlines in multiple input files with pdfwrite
Bug #705218 "table of contents/bookmarks wrong after merging PDF files"

Part 2 of 2 for this bug.

The pdfwrite device is capable of processing pdfmarks to produce a
variety of effects in the output. In this case we can add Link
annotations and Outlines to a PDF file. When the PDF interpreter
supports reading these from an input file we also write them to the
output file using (effectively) pdfmark operationjs.

The problem is that when we have multiple input files and the Dest of
a Link annotation or Outline is a page, that page refers to the page
number in the original file (eg 1). That won't be the correct page
number in the output file because it already contains the pages from
any previous files.

So what we need to do is offset the page number of the destination by
the number of pages already in the output. That's what this commit does;
for Ghostscript processing we store the number of pages in the device
and update at the end of each file by the number of pages in that file.
We send the number of pages to the (newly created) interpreter at the
start of every file in the dictionary argument we supply to .PDFInit.

For GhostPDF it is simpler because we handle all the files, we just
update the counter in the PDF context by the number of pages processed
in the file we have just completed.

We simply add the offset to the page number when creating the Dest
pdfmark.

This commit also extends the named destination processing to handle
name trees with more than a single node by recursively processing the
Kids array in each node. We also use the Limits to more quickly
determine if a node contains our target, rather than checking each
entry in the Names array for a match.