GhostPDF - pad uneven hex strings with 0
Bug #705396 "Error with short hex strings"

The PDF Reference (v 1.7) page 56 says "If the final digit of a
hexadecimal string is missing—that is, if there is an odd number of
digits—the final digit is assumed to be 0."

We weren't doing that causing an error with the file. The Panose strings
are invalid even when that is taken into account (don't make 12 bytes
that way) but at least we shouldn't throw an error.

This commit simply follows what the spec says, if we don't get an even
number of bytes then assume the last byte is 0.