Fix overprint_copy_planes with regions of height > 1.
The routine loops 1 scanline at a time, reading a line of
background data, overwriting it if required, and then
writing it back.

The decision as to whether it should be overwritten or not
is based upon bits in comps. Unfortunately, as the loop proceeds
it destroys comps, so it's only valid for the first scanline.

Here we restore the original value of comps each time around
the y loop, so we operate correctly.

No reproducer file - spotted while investigating bug 705260.