Bug 705260: Add gxdso for setting overprint operation.
The overprint compositor needs to know whether we are doing a
fill or a stroke to know what to do. Usually this is achieved
by the overprint device intercepting the fill/stroke operations
and setting the state before passing on.

Unfortunately, for 'fill_stroke' operations that aren't handled
by the overprint device, we have no way of knowing whether we are
filling or stroking.

The case of this shown by the bug is when we have alpha bits in
play. We enter the graphics lib to do a fill stroke, with the
erasepage_optimisation device in play (the erasepage stuff is not
important, it could be the base device too). We wrap that in an
overprint device, and then (because we have alphabits), we wrap
that in the alphabits device.

When we call the fill_stroke_path device proc, the alphabits device
forwards this to gx_default_fill_stroke path, which does a fill then
a stroke, without the overprint compositor being told what operation
is what.

If we were to make the alphabits device simply forward to the underlying
device (the overprint device), we'd correctly set fill/stroke as
appropriate, but would end up writing 8 bits color into the alphabits
device (which is greyscale, and typically 4 bits).

Accordingly, we implement a new function for the alphabits device
to handle fill_stroke_path that does pretty much what the default
fill_stroke_path would do, but also fires off a new devspecop to
inform any overprint compositor of the change in state required.