More repair work on PDFDEBUG
We can't just have the keywords echoed when interpreted by the stream
interpretation, because then they don't get output for file repair.

Move the dumping of keywords from pdfi_interpret_stream_operator() into
the keyword parsing. This means we need it in two places now; one for
'bare' keywords and one for the full-blown keyword parsing.