GhostPDF - repair files with extra invalid xref entries
Bug #705428 "Broken PDF with xref larger than number of objects defined in file can't be repaired"

If the PDF was broken enough to trigger a repair then the repair code
would duly rescan the file, filling in the xref with the offsets of all
the objects it found. If there are existing xref entries, which we did
not find when scanning the file, we still examine those trying to
determine if they are ObjStm types.

If the offset to the object does not point to a valid object (which it
very likely wouldn't) then we errored out of repairing the file, which
causes the interpreter to give up.

This commit ignores any object which fails to parse when checked which
should cause an error if we ever try to actually use it, but doesn't
cause the repair to fail. Since the file is already broken enough to
require repairing this shouldn't be a surprise to anyone.

Additionally, if the offset points past the end of file, treat it the
same way.