Bug 695153: Tweak interpolation edge case.
The example file (or at least the cut down one), contains a smallish
(256x256) image that is scaled up a lot (to 35379449,35379449). This
does not fit on the 360x240 output (at 10dpi), so only a small
proportion is required.

Because interpolation is required, we don't want to calculate all
1.25 quadrillion output pixels. Accordingly, we calculate the
'patch' region of the source that is actually required, and then
only process that region.

Make the calculations more accurate to try to minimise the
source region. In particular, on an upscale, it doesn't need to be
increased as much as on a downscale.