GhostPDF - prevent infinite loop on token parsing error
This exhibited with the file for the OSS-fuzz bug #42299, if we got an
invalid token after finding an 'endstream' (ie the endobj was missing)
then we could end up in an infinite loop trying to find a token and
failing.

Deal with that by treating an invalid key (after finding an endstream)
as if we had found an endobj.