Subclassing devices - optionally keep device before compositor
When we push the transparency compositor there are times when we would
like to keep an existing device in front of the transparency compositor,
in particular the object filtering device subclasses.

This is because after the compositor all operations are an image which
makes filtering on the object type impossible.

Currently the subclassing code goes to great lengths to keep the
subclassing device tightly bound to the subclassed device, patching up
the compositor so that it references the subclassing device rather
than the base device.

This commit adds a new compositor function for subclassing devices
'default_subclass_compositor_front' which starts off by calling the
child's compositor method. If that returns '1' (we created a new
compositor and want it to be the current device) then we point the
subclassing device child member at the returned device instead of the
subclassed device. We remember the device we were pointing at though as
we will need this later. We then patch up the compositor's 'finalize'
method, storing the one it starts with and replacing it with one of our
own.

Finally we return '0' so that the caller thinks there was no need to
create a new compositor and so does not attempt to make it the current
device in the graphics state.

If the compositor is freed (end of page/job restore) our custom
finalize procedure gets called. That calls the correct, stored, finalize
for the compositor, and points the subclassing device's child member
back to the subclassed device.

However we also process compositor 'pop' operations, and in that case
we will remove the compositor from the device chanin and count it down
(thus freeing it). Obviously we need to be careful to restore the
original finalize method before doing this.

In effect we insert the compositor and remove it leaving the chain of
devices unchanged afterward.

Now the transparency compositor can change its colour information,
presumably based on whatever blending space the current group is using.
This was causing problems when we put a subclass device ahead of the
compositor, as the subclassing device wasn't changing it's colour
information when the transparency compositor did.

So we must now force the colour info of the subclassing device to match
the current colour info of the compositor after the compositor has
processed the operation. This means we need to process all tranparency
actions, though apart from push and pop we only synchronise the colour
info.

We use the original child device to restore the colour info of the
subclassing device when we pop the compositor.

This (finally!) results in no seg faults on the cluster when run with
extras=-dFILTERIMAGE though it does cause one file (Bug692217.pdf) to
error when run with the pdfwrite device. The error occurs reading the
pdfwrite output file, in the transparency code, and I've opened a bug
for it as the error occurs when reading the file without FILTERIMAGE
see bug 705519

A subclassing device which wants to stay ahead of the transparency
compositor must set its 'composite' method to be the subclass
'default_subclass_composite_front' method rather than the standard
'default_subclass_composite' method.

There is still (I think) a potential problem with patterns. The
pattern accumulator device is pushed to the current device when
a pattern starts, and there is currently no way to intercept that
actio. This means that objects which would otherwise be filtered
out will be stored into the pattern. I'm not certain if rendering the
pattern, and then reusing the bitmap, will result in the correct
output or not.