Object filtering - support image type 3 and 3x
Bug # 705366 "Problem deleting images in a PDF"

The bug report says that Ghostscript 'fails and freezes', which I do
not see. What I do see is that the image in the file, when run with
-dFILTERIMAGE and pdfwrite, generates an error.

This is because the object filtering code does not support image types
3 and 3x (a GS special type for images with SMasks). This has always
caused an error, which the old interpreter silently ignores.

This commit adds decent support for reading and discarding the correct
amount of data from the image source(s), so that there is no longer an
error. The output is unchanged.