SUBSTFONT functionality
This was missing from pdfi, so is added here.

Also, change the functionality so setting the SUBSTFONT key to "/None" will
result in any attempt to fall back to the default font to throw an
invalidfont error.

Lastly, fix some typos on the parameters being passed from Postscript into
pdfi.