OSS-fuzz #48111
Before objects could be low value integers, and were always pointers
to a structure, it was acceptable to modify the members of the
structure before checking the object type.

Doing so now can result in a seg fault if the object isn't a pointer to
a structure.

So move the assignment of the indirect object number past the point
where we check the object type to ensure that the object is indeed a
pointer to a structure and dereferencing it is safe.