OSS-fuzz #48128 - limit frequency in halftone setup
We use an intermediate variable 'rt', and we calculate its value by
running round a loop, incrementing it by one, until '(u+v) * rt' is at
least 4.

However, rt is an integer and u and v are both doubles; this means that
they can be so small that multiplying by rt never reaches 4, resulting
in an infinite loop.

This commit simply adds a check to ensure that the loop will terminate
and returns a rangecheck if it will not. On Windows this will allow
for frequency values up to 100 million lpi, on Linux up to 1x10^19 lpi
(Linux has a 64-bit integer, Windows has 32). I think this is large
enough for any reasonable halftone screen.