oss-fuzz 47770: Init null device on creation
Under normal (rendering) operation, for PDF tr 3 text, we send the text
operation to a null device, meaning the graphics gets updated properly, without
actually rendering anything on the output.

Previously code would allocated the null device, do a gs_gsave() and the init
and set the null device.

The problem is that, if the gsave fails, we leave an unitialised device for
the garbager to clean up, causing a crash.

Secondly, any interpreter not using the garbager would leak memory in that
error condition.

So, move the device initialisation before the gs_gsave() call, and explicitly
free the device in the event of an error.