pdfwrite - change the way some images are written
This is from a customer query; type 4 (chroma-keyed) images, when
colour converted, required the code to be permitted to select a
compression strategy automatically. Disabling the auto filtering
caused an error (deliberately; indicating an incompatible combination
of options).

This commit modifies the way some manufactured images are stored so that
we pass the 'for_pattern' parameter, which is used by
pdf_copy_color_data, from a much higher level, allowing us to
discriminate between various types of 'colour image' more precisely.

This is needed because the same code is used to write the image from
converting a type 4 image, writing the image resulting from rendering a
shading pattern and writing the rendered image from a tiling pattern.
And possibly other places as yet unidentified.....

This commit may be reverted in future if it causes problems.