OSS-fuzz #48167
This comes about because of the rework to return integers directly
instead of as pointers to objects. The object mechanism pushed the
objects onto the stack and so, on an error, popped them from the stack.

The new method doesn't use the stack, so we must not pop objects from
it! Even on error....