GhostPDF - don't reuse stack references
OSS-fuzz bug 48167 demonstrated that there are still places where we
reference an object on the stack, but do not count it up and expect it
to remain valid.

This is not a safe assumption. Triggering a repair can clear the stack,
and in the case of the bug, a code error can lead to the stack object
being removed.

This commit goes through all the places that we seem to make that
assumption and reworks them to instead take an explicit reference.