GhostPDF - more work on Info dictionaries....
OSS-fuzz #48205

Now that objects are not always pointers to structures, we can't assume
that we can simply dereference them.

Fortunately the only objects for which we want to store the object
number in this case are the composite objects (array and dictionary) so
we can just store the object number for those.