GhostPDF - more strange Info dictionaries
The AirPrint test file has a Info dictionary where the CreationDtae and
ModDate both reference the same indirect object. This was causing the
circular reference detection to think we had a circular reference.

To avoid this, at the top level clear the circular reference stack after
each fetch from the dictionary.