OSS-fuzz #48283 - limit exponent numbers
The actual timeout is caused because the file is very badly corrupted
and contains a slew of numbers which have the form <nu>e<exp>, and in
some cases the exponent is exceedingly large (> 2500).

This causes sscanf to return a #INF, and the graphics library turns that
into the largest possible number. The upshot is that we end up with a
curve from nearly 0 to somewhere in the next country. Which we then
need to flatten. Needless to say this takes an exceedingly long time.

Although the processing time is legitimate, and we could potentially
get here with a perfectly valid path which would take a very long time
to process, in this case the problem is caused by the illegal number.

So add some checking to the exponent number case (Note: exponent format
numbers are specifically prohibited in the spec, but Acrobat is known
to accept them....) to treat values > e38 as 0, because that's the
largest real number Acrobat will accept. This shouldn't be a real
performance bottleneck because we should never see a exponent number
so they ought to be rare. Obviously we raise an error for this as well.

This makes the path much more sensible and completes in a reasonable
time.